put db client functions all into one file
